login function and refactory
this is a first version we need to wait for backend and redux to be ready to finish
